fix(library): address code review security and platform findings
