+++ library/payments/qbo/internal/cli/helpers.go
@@ -4,27 +4,16 @@
 package cli
 
 import (
-	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"github.com/mvanhorn/printing-press-library/library/payments/qbo/internal/client"
-	"github.com/mvanhorn/printing-press-library/library/payments/qbo/internal/cliutil"
 	"github.com/spf13/cobra"
-	"github.com/spf13/pflag"
 	"io"
 	"net/url"
 	"os"
 	"sort"
-	"strconv"
 	"strings"
 	"text/tabwriter"
-	"unicode"
 )
-
-var As = errors.As
-
-const paginatedGetMaxPages = 100
 
 // noColor is set by the --no-color flag
 var noColor bool
@@ -36,175 +25,55 @@ func colorEnabled() bool {
 	if noColor {
 		return false
 	}
-	if !humanFriendly {
-		return false
-	}
-	if os.Getenv("NO_COLOR") != "" {
-		return false
-	}
-	if os.Getenv("TERM") == "dumb" {
-		return false
-	}
 	return isTerminal(os.Stdout)
 }
 
 func isTerminal(w io.Writer) bool {
 	if f, ok := w.(*os.File); ok {
-		fi, err := f.Stat()
-		if err != nil {
-			return true
+		stat, err := f.Stat()
+		if err == nil {
+			return (stat.Mode() & os.ModeCharDevice) != 0
 		}
-		return (fi.Mode() & os.ModeCharDevice) != 0
 	}
 	return false
 }
 
 func bold(s string) string {
-	if !colorEnabled() {
-		return s
+	if colorEnabled() {
+		return "\x1b[1m" + s + "\x1b[0m"
 	}
-	return "\033[1m" + s + "\033[0m"
+	return s
 }
 
 func green(s string) string {
-	if !colorEnabled() {
-		return s
+	if colorEnabled() {
+		return "\x1b[32m" + s + "\x1b[0m"
 	}
-	return "\033[32m" + s + "\033[0m"
+	return s
 }
 
 func red(s string) string {
-	if !colorEnabled() {
-		return s
+	if colorEnabled() {
+		return "\x1b[31m" + s + "\x1b[0m"
 	}
-	return "\033[31m" + s + "\033[0m"
+	return s
 }
 
 func yellow(s string) string {
-	if !colorEnabled() {
-		return s
+	if colorEnabled() {
+		return "\x1b[33m" + s + "\x1b[0m"
 	}
-	return "\033[33m" + s + "\033[0m"
-}
-
-type cliError struct {
-	code int
-	err  error
-}
-
-func (e *cliError) Error() string { return e.err.Error() }
-func (e *cliError) Unwrap() error { return e.err }
-
-func usageErr(err error) error     { return &cliError{code: 2, err: err} }
-func notFoundErr(err error) error  { return &cliError{code: 3, err: err} }
-func authErr(err error) error      { return &cliError{code: 4, err: err} }
-func apiErr(err error) error       { return &cliError{code: 5, err: err} }
-func configErr(err error) error    { return &cliError{code: 10, err: err} }
-func rateLimitErr(err error) error { return &cliError{code: 7, err: err} }
-
-// partialFailureErr signals that the upstream API returned a 2xx with a
-// body shape indicating some operations in a batch failed (e.g. Google
-// Ads `partialFailureError`, similar shapes from Drive batch, Sheets
-// batchUpdate, Cloud Resource Manager). Distinct from apiErr (HTTP-level
-// failure) so callers can distinguish "request rejected" from "request
-// accepted but some ops failed".
-func partialFailureErr(err error) error { return &cliError{code: 6, err: err} }
-
-// partialFailureReport describes the structured detection result for a
-// mutate-style response body. Emitted in the envelope under
-// "partial_failure" so machine-readable callers can route per-operation
-// remediation.
-type partialFailureReport struct {
-	Field         string   `json:"field"`
-	Message       string   `json:"message,omitempty"`
-	Code          int      `json:"code,omitempty"`
-	Details       any      `json:"details,omitempty"`
-	ResourceNames []string `json:"resource_names,omitempty"`
-}
-
-// detectPartialFailure inspects a mutate-style JSON response for a
-// partial-failure-shaped field. Returns nil when no partial failure is
-// detected. The detector is intentionally generic across APIs that emit
-// 2xx-with-batch-errors. New partial-failure-shaped fields are added to
-// partialFailureFields, not at call sites. When `results[]` is present
-// (Google Ads convention) it extracts per-op `resourceName` so callers
-// can see which operations did succeed.
-func detectPartialFailure(data []byte) *partialFailureReport {
-	if len(data) == 0 {
-		return nil
-	}
-	var top map[string]any
-	if err := json.Unmarshal(data, &top); err != nil {
-		return nil
-	}
-	partialFailureFields := []string{"partialFailureError"}
-	for _, field := range partialFailureFields {
-		raw, ok := top[field]
-		if !ok || raw == nil {
-			continue
-		}
-		obj, ok := raw.(map[string]any)
-		if !ok {
-			continue
-		}
-		message, _ := obj["message"].(string)
-		var code int
-		if n, ok := obj["code"].(float64); ok {
-			code = int(n)
-		}
-		// Empty object means partial-failure mode was off or no ops
-		// failed; do not flag.
-		if code == 0 && strings.TrimSpace(message) == "" {
-			continue
-		}
-		report := &partialFailureReport{
-			Field:   field,
-			Message: message,
-			Code:    code,
-			Details: obj["details"],
-		}
-		if results, ok := top["results"].([]any); ok {
-			for _, r := range results {
-				if rm, ok := r.(map[string]any); ok {
-					if name, ok := rm["resourceName"].(string); ok && name != "" {
-						report.ResourceNames = append(report.ResourceNames, name)
-					}
-				}
-			}
-		}
-		return report
-	}
-	return nil
+	return s
 }
 
 // dryRunOK reports whether the command should short-circuit without doing any
-// real work because --dry-run was set. The verify pipeline probes hand-written
-// commands with --dry-run; commands that put validation in cobra's `Args:` or
-// `MarkFlagRequired` cannot reach a dry-run guard inside RunE because cobra
-// runs those checks before RunE. The verify-friendly pattern for hand-written
-// commands is:
-//
-//	RunE: func(cmd *cobra.Command, args []string) error {
-//	    if len(args) == 0 {
-//	        return cmd.Help()
-//	    }
-//	    if dryRunOK(flags) {
-//	        return nil
-//	    }
-//	    // ... real work ...
-//	}
-//
-// See SKILL.md "Phase 3: Build The GOAT" for the full pattern.
+// real work because --dry-run was set.
 func dryRunOK(flags *rootFlags) bool {
 	return flags != nil && flags.dryRun
 }
 
 // parentNoSubcommandRunE returns a RunE that handles parents invoked without a
-// subcommand. In machine output (--json/--agent) the parent emits a structured
-// error to stdout listing valid subcommands and exits 2; otherwise cobra's
-// default help text is printed. Without this, agents driving the CLI in
-// --agent mode received only human-readable help on stdout and exit 0, with no
-// signal that the invocation was incomplete.
+// subcommand.
 func parentNoSubcommandRunE(flags *rootFlags) func(*cobra.Command, []string) error {
 	return func(cmd *cobra.Command, args []string) error {
 		if flags != nil && flags.asJSON {
@@ -238,57 +107,6 @@ func writeNoop(flags *rootFlags, reason, prose string) error {
 	return nil
 }
 
-func writeAPIErrorEnvelope(flags *rootFlags, err error, code int) {
-	if flags == nil || !flags.asJSON {
-		return
-	}
-	_ = json.NewEncoder(os.Stdout).Encode(map[string]any{
-		"error": err.Error(),
-		"code":  code,
-	})
-}
-
-// classifyAPIError maps API errors to structured exit codes with actionable hints.
-func classifyAPIError(err error, flags *rootFlags) error {
-	var typed *cliError
-	if errors.As(err, &typed) {
-		return err
-	}
-
-	msg := err.Error()
-	switch {
-	case strings.Contains(msg, "HTTP 409"):
-		if flags != nil && flags.idempotent {
-			return writeNoop(flags, "already_exists", "already exists (no-op)")
-		}
-		classified := apiErr(err)
-		writeAPIErrorEnvelope(flags, classified, ExitCode(classified))
-		return classified
-	case errors.Is(err, client.ErrPlaceholderCredential):
-		return authErr(err)
-	case strings.Contains(msg, "HTTP 400") && cliutil.LooksLikeAuthError(msg):
-		return authErr(fmt.Errorf("%w\nhint: the API rejected the request — this usually means auth is missing or invalid."+
-			"\n      Set your API key: export QBO_CLIENT_ID=<your-key>"+
-			"\n      Run 'qbo-pp-cli doctor' to check auth status."+
-			"\n      Response: "+cliutil.SanitizeErrorBody(msg), err))
-	case strings.Contains(msg, "HTTP 401"):
-		return authErr(fmt.Errorf("%w\nhint: check your token. Set it with: qbo-pp-cli auth set-token <token>"+
-			"\n      or: export QBO_CLIENT_ID=<your-token>"+
-			"\n      Run 'qbo-pp-cli doctor' to check auth status.", err))
-	case strings.Contains(msg, "HTTP 403"):
-		return authErr(fmt.Errorf("%w\nhint: permission denied. Your credentials are valid but lack access to this resource."+
-			"\n      Check that your API key has the required permissions."+
-			"\n      Set it with: export QBO_CLIENT_ID=<your-key>"+
-			"\n      Run 'qbo-pp-cli doctor' to check auth status.", err))
-	case strings.Contains(msg, "HTTP 404"):
-		return notFoundErr(fmt.Errorf("%w\nhint: resource not found. Run the 'list' command to see available items", err))
-	case strings.Contains(msg, "HTTP 429"):
-		return rateLimitErr(err)
-	default:
-		return apiErr(err)
-	}
-}
-
 func truncate(s string, max int) string {
 	if len(s) <= max {
 		return s
@@ -307,1139 +125,4 @@ func newTabWriter(w io.Writer) *tabwriter.Writer {
 // user input do not collapse into extra path segments.
 func replacePathParam(path, name, value string) string {
 	return strings.ReplaceAll(path, "{"+name+"}", url.PathEscape(value))
-}
-
-// paginatedGet fetches pages and concatenates array results. The headers
-// argument carries per-endpoint required headers (e.g. cal-api-version) that
-// must be sent on every page request, including the first; pass nil when the
-// endpoint has no per-endpoint header overrides.
-func paginatedGet(ctx context.Context, c interface {
-	GetWithHeaders(ctx context.Context, path string, params map[string]string, headers map[string]string) (json.RawMessage, error)
-}, path string, params map[string]string, headers map[string]string, fetchAll bool, cursorParam, paginationType, limitParam, nextCursorPath, hasMoreField string) (json.RawMessage, error) {
-	// Cursor params are exempt from the "0"/"false" strip: offset-paginated
-	// APIs send offset=0 on the first page.
-	clean := map[string]string{}
-	for k, v := range params {
-		if v == "" {
-			continue
-		}
-		if k == cursorParam || (v != "0" && v != "false") {
-			clean[k] = v
-		}
-	}
-
-	if !fetchAll {
-		data, err := c.GetWithHeaders(ctx, path, clean, headers)
-		if err != nil {
-			return nil, err
-		}
-		emitTruncationWarning(data, nextCursorPath, hasMoreField, paginationType)
-		return data, nil
-	}
-
-	// Fetch all pages
-	allItems := make([]json.RawMessage, 0)
-	page := 0
-	for {
-		page++
-		if humanFriendly {
-			fmt.Fprintf(os.Stderr, "fetching page %d...\n", page)
-		} else {
-			fmt.Fprintf(os.Stderr, `{"event":"page_fetch","page":%d}`+"\n", page)
-		}
-
-		data, err := c.GetWithHeaders(ctx, path, clean, headers)
-		if err != nil {
-			return nil, err
-		}
-
-		// Try to extract items array
-		var items []json.RawMessage
-		if json.Unmarshal(data, &items) == nil {
-			allItems = append(allItems, items...)
-		} else {
-			// Response is an object - look for array inside
-			var obj map[string]json.RawMessage
-			if json.Unmarshal(data, &obj) == nil {
-				if nested, ok := extractPaginatedItems(obj); ok {
-					allItems = append(allItems, nested...)
-				}
-
-				// Check for next cursor
-				if nextCursorPath != "" {
-					if tokenRaw, ok := rawAtPath(obj, nextCursorPath); ok {
-						if token := paginationCursorToken(tokenRaw); token != "" {
-							if page >= paginatedGetMaxPages {
-								emitPaginatedGetMaxPagesWarning()
-								break
-							}
-							clean[cursorParam] = token
-							continue
-						}
-					}
-				}
-
-				// Check has_more. Page and offset paginators can advance
-				// client-side; cursor-based APIs still need a body cursor.
-				if hasMoreField != "" {
-					if moreRaw, ok := rawAtPath(obj, hasMoreField); ok {
-						var more bool
-						if json.Unmarshal(moreRaw, &more) == nil && more {
-							if next, ok := nextClientSidePaginationCursor(clean, cursorParam, paginationType, limitParam); ok {
-								if page >= paginatedGetMaxPages {
-									emitPaginatedGetMaxPagesWarning()
-									break
-								}
-								clean[cursorParam] = next
-								continue
-							}
-							emitMissingPaginationCursorWarning(nextCursorPath)
-							break
-						}
-					}
-				}
-			}
-			// No more pages
-			break
-		}
-
-		// For direct arrays, can't paginate without cursor
-		break
-	}
-
-	if fetchAll && page == 1 && nextCursorPath == "" && hasMoreField == "" {
-		emitMissingPaginationSignalWarning()
-	}
-	if humanFriendly {
-		fmt.Fprintf(os.Stderr, "fetched %d items across %d pages\n", len(allItems), page)
-	} else {
-		fmt.Fprintf(os.Stderr, `{"event":"complete","total":%d,"pages":%d}`+"\n", len(allItems), page)
-	}
-	result, _ := json.Marshal(allItems)
-	return json.RawMessage(result), nil
-}
-
-func nextClientSidePaginationCursor(params map[string]string, cursorParam, paginationType, limitParam string) (string, bool) {
-	if cursorParam == "" {
-		return "", false
-	}
-	switch paginationType {
-	case "page":
-		current := params[cursorParam]
-		if current == "" {
-			current = "1"
-		}
-		n, err := strconv.Atoi(current)
-		if err != nil {
-			return "", false
-		}
-		return strconv.Itoa(n + 1), true
-	case "offset":
-		current := params[cursorParam]
-		if current == "" {
-			current = "0"
-		}
-		n, err := strconv.Atoi(current)
-		if err != nil {
-			return "", false
-		}
-		limit, err := strconv.Atoi(params[limitParam])
-		if err != nil || limit <= 0 {
-			return "", false
-		}
-		return strconv.Itoa(n + limit), true
-	default:
-		return "", false
-	}
-}
-
-// Silent page-1 truncation is the worst-possible mode for agents,
-// who otherwise compute totals against an incomplete set without
-// passing --all.
-func emitTruncationWarning(data json.RawMessage, nextCursorPath, hasMoreField, paginationType string) {
-	if nextCursorPath == "" && hasMoreField == "" {
-		return
-	}
-	var obj map[string]json.RawMessage
-	if err := json.Unmarshal(data, &obj); err != nil {
-		return
-	}
-	var nextCursor string
-	if nextCursorPath != "" {
-		if tokenRaw, ok := rawAtPath(obj, nextCursorPath); ok {
-			nextCursor = paginationCursorToken(tokenRaw)
-		}
-	}
-	var hasMore bool
-	if hasMoreField != "" {
-		if moreRaw, ok := rawAtPath(obj, hasMoreField); ok {
-			_ = json.Unmarshal(moreRaw, &hasMore)
-		}
-	}
-	if nextCursor == "" && !hasMore {
-		return
-	}
-	// --all advances when a next-cursor is configured, or when the endpoint
-	// uses client-side numeric page/offset advancement. Opaque cursor APIs
-	// still need a returned cursor to advance safely.
-	if nextCursor != "" || ((paginationType == "page" || paginationType == "offset") && hasMore) {
-		if humanFriendly {
-			fmt.Fprintf(os.Stderr, "warning: results truncated; more pages available. Re-run with --all to fetch every page.\n")
-		} else {
-			fmt.Fprintf(os.Stderr, `{"event":"truncated","hint":"pass --all to fetch every page"}`+"\n")
-		}
-		return
-	}
-	if humanFriendly {
-		fmt.Fprintf(os.Stderr, "warning: results truncated; more pages available.\n")
-	} else {
-		fmt.Fprintf(os.Stderr, `{"event":"truncated"}`+"\n")
-	}
-}
-
-func emitMissingPaginationSignalWarning() {
-	if humanFriendly {
-		fmt.Fprintf(os.Stderr, "warning: --all requested, but this endpoint does not declare a next cursor or has-more field; returning page 1 only.\n")
-	} else {
-		fmt.Fprintf(os.Stderr, `{"event":"truncated","reason":"pagination_signal_missing","message":"--all requested but this endpoint does not declare a next cursor or has-more field; returning page 1 only"}`+"\n")
-	}
-}
-
-func emitPaginatedGetMaxPagesWarning() {
-	if humanFriendly {
-		fmt.Fprintf(os.Stderr, "warning: --all reached the %d-page safety limit; returning fetched pages only.\n", paginatedGetMaxPages)
-	} else {
-		fmt.Fprintf(os.Stderr, `{"event":"truncated","reason":"max_pages_cap_hit","message":"--all reached the %d-page safety limit; returning fetched pages only"}`+"\n", paginatedGetMaxPages)
-	}
-}
-
-func emitMissingPaginationCursorWarning(nextCursorPath string) {
-	if humanFriendly {
-		fmt.Fprintf(os.Stderr, "warning: --all requested, but the response indicated more pages without a usable next cursor; returning fetched pages only.\n")
-	} else if nextCursorPath != "" {
-		fmt.Fprintf(os.Stderr, `{"event":"truncated","reason":"pagination_cursor_missing","next_cursor_path":%q,"message":"--all requested but the response indicated more pages without a usable next cursor; returning fetched pages only"}`+"\n", nextCursorPath)
-	} else {
-		fmt.Fprintf(os.Stderr, `{"event":"truncated","reason":"pagination_cursor_missing","message":"--all requested but the response indicated more pages without a usable next cursor; returning fetched pages only"}`+"\n")
-	}
-}
-
-func paginationCursorToken(raw json.RawMessage) string {
-	var token string
-	if json.Unmarshal(raw, &token) == nil && token != "" {
-		return token
-	}
-	var number json.Number
-	if json.Unmarshal(raw, &number) == nil {
-		if n, err := number.Int64(); err == nil && n > 0 {
-			return number.String()
-		}
-	}
-	return ""
-}
-
-func extractPaginatedItems(obj map[string]json.RawMessage) ([]json.RawMessage, bool) {
-	for _, field := range []string{"data", "items", "results", "messages", "members", "values"} {
-		if arr, ok := obj[field]; ok {
-			var nested []json.RawMessage
-			if json.Unmarshal(arr, &nested) == nil {
-				return nested, true
-			}
-		}
-	}
-
-	var onlyArray []json.RawMessage
-	arrayCount := 0
-	for _, raw := range obj {
-		var candidate []json.RawMessage
-		if json.Unmarshal(raw, &candidate) == nil {
-			onlyArray = candidate
-			arrayCount++
-		}
-	}
-	if arrayCount == 1 {
-		return onlyArray, true
-	}
-	return nil, false
-}
-
-func rawAtPath(obj map[string]json.RawMessage, path string) (json.RawMessage, bool) {
-	if raw, ok := obj[path]; ok {
-		return raw, true
-	}
-
-	current := obj
-	parts := strings.Split(path, ".")
-	for i, part := range parts {
-		raw, ok := current[part]
-		if !ok {
-			return nil, false
-		}
-		if i == len(parts)-1 {
-			return raw, true
-		}
-		if err := json.Unmarshal(raw, &current); err != nil {
-			return nil, false
-		}
-	}
-	return nil, false
-}
-
-// printJSONFiltered marshals a Go-typed value through the same output
-// pipeline endpoint-mirror commands use. Hand-written novel commands that
-// build a typed slice/struct call this so --select, --compact, --csv, and
-// --quiet all behave the same way as on generator-emitted commands.
-func printJSONFiltered(w io.Writer, v any, flags *rootFlags) error {
-	raw, err := json.Marshal(v)
-	if err != nil {
-		return err
-	}
-	return printOutputWithFlags(w, json.RawMessage(raw), flags)
-}
-
-// filterFields keeps only the specified fields (comma-separated) from JSON objects/arrays.
-// Supports dotted paths like "events.shortName" to descend into nested structures.
-// Arrays are traversed element-wise: "events.shortName" keeps shortName on each event.
-func filterFields(data json.RawMessage, fields string) json.RawMessage {
-	var paths [][]string
-	for _, f := range strings.Split(fields, ",") {
-		f = strings.TrimSpace(f)
-		if f == "" {
-			continue
-		}
-		parts := strings.Split(f, ".")
-		for i := range parts {
-			parts[i] = strings.ToLower(parts[i])
-		}
-		paths = append(paths, parts)
-	}
-	if len(paths) == 0 {
-		return data
-	}
-	return filterFieldsRec(data, paths)
-}
-
-// filterFieldsRec applies path filters to a JSON value. Each path is a list of
-// lowercase segments; arrays descend element-wise.
-func filterFieldsRec(data json.RawMessage, paths [][]string) json.RawMessage {
-	var arr []json.RawMessage
-	if err := json.Unmarshal(data, &arr); err == nil {
-		out := make([]json.RawMessage, len(arr))
-		for i, el := range arr {
-			out[i] = filterFieldsRec(el, paths)
-		}
-		result, _ := json.Marshal(out)
-		return result
-	}
-
-	var obj map[string]json.RawMessage
-	if err := json.Unmarshal(data, &obj); err == nil {
-		keepWhole := map[string]bool{}
-		subPaths := map[string][][]string{}
-		for _, p := range paths {
-			if len(p) == 0 {
-				continue
-			}
-			head := p[0]
-			if len(p) == 1 {
-				keepWhole[head] = true
-			} else {
-				subPaths[head] = append(subPaths[head], p[1:])
-			}
-		}
-		filtered := map[string]json.RawMessage{}
-		matchedAny := false
-		for k, v := range obj {
-			matched := matchSelectSegment(k, keepWhole, subPaths)
-			if matched == "" {
-				continue
-			}
-			matchedAny = true
-			if keepWhole[matched] {
-				filtered[k] = v
-				continue
-			}
-			if subs := subPaths[matched]; subs != nil {
-				filtered[k] = filterFieldsRec(v, subs)
-			}
-		}
-		// Envelope fallback: when no top-level keys matched but at least one
-		// sibling is a non-null array, treat the object as a list envelope
-		// (`{"items":[...]}`, `{"data":[...]}`, `{"total_count":N,"items":[...]}`)
-		// and apply the selector inside the array(s). Non-array siblings pass
-		// through verbatim so envelope metadata (counts, null pagination
-		// cursors) stays visible. The foundArray guard preserves the prior
-		// empty-object result for flat objects where no key matches and no
-		// array exists. The `arr != nil` check rejects JSON null, which
-		// json.Unmarshal otherwise accepts into a []json.RawMessage as a
-		// nil slice and would coerce to `[]`.
-		if !matchedAny {
-			pending := map[string]json.RawMessage{}
-			foundArray := false
-			for k, v := range obj {
-				var arr []json.RawMessage
-				if json.Unmarshal(v, &arr) == nil && arr != nil {
-					foundArray = true
-					pending[k] = filterFieldsRec(v, paths)
-				} else {
-					pending[k] = v
-				}
-			}
-			if foundArray {
-				for k, v := range pending {
-					filtered[k] = v
-				}
-			}
-		}
-		result, _ := json.Marshal(filtered)
-		return result
-	}
-
-	return data
-}
-
-// matchSelectSegment returns the matching lowercase segment, or "" if no match.
-// Supports direct case-insensitive match and camelCase→kebab-case conversion.
-func matchSelectSegment(fieldName string, keepWhole map[string]bool, subPaths map[string][][]string) string {
-	lower := strings.ToLower(fieldName)
-	if keepWhole[lower] || subPaths[lower] != nil {
-		return lower
-	}
-	kebab := camelToKebab(fieldName)
-	if kebab != lower && (keepWhole[kebab] || subPaths[kebab] != nil) {
-		return kebab
-	}
-	return ""
-}
-
-// camelToKebab converts "orderDate" or "orderdate" to "order-date" by splitting on
-// uppercase boundaries. For already-lowercase input, splits on known word boundaries.
-func camelToKebab(s string) string {
-	var b strings.Builder
-	runes := []rune(s)
-	for i, r := range runes {
-		if i > 0 && unicode.IsUpper(r) && unicode.IsLower(runes[i-1]) {
-			b.WriteByte('-')
-		}
-		b.WriteRune(unicode.ToLower(r))
-	}
-	return b.String()
-}
-
-// printOutputWithFlags routes output through the right format based on flags.
-func printOutputWithFlags(w io.Writer, data json.RawMessage, flags *rootFlags) error {
-	// --select wins over --compact when both are set: an explicit field list
-	// is the user's authoritative request, so the high-gravity allow-list
-	// must not strip those fields out before --select can pick them. When
-	// only --compact is set (e.g., --agent without --select), the allow-list
-	// still runs.
-	if flags.selectFields != "" {
-		data = filterFields(data, flags.selectFields)
-	} else if flags.compact {
-		data = compactFields(data)
-	}
-	// --quiet: suppress all output, exit code communicates result
-	if flags.quiet {
-		return nil
-	}
-	// --csv: render as CSV
-	if flags.csv {
-		return printCSV(w, data)
-	}
-	return printOutput(w, data, flags.asJSON)
-}
-
-// compactVerboseListFields are prose-shaped fields stripped from list-item
-// projections. On lists, "body"/"content"/"html"/"markdown" are verbose
-// noise and the row's identity is carried by id/name/title/etc.
-var compactVerboseListFields = map[string]bool{
-	"description": true, "body": true, "content": true,
-	"comments": true, "attachments": true, "html": true, "markdown": true,
-}
-
-// compactVerboseObjectFields are metadata fields stripped from single-object
-// responses. "body"/"content"/"html"/"markdown" are intentionally absent:
-// for a `get` command those fields are the primary payload, and stripping
-// them under `--agent`/`--compact` silently emits a useless envelope.
-// Use `--select` to drop them explicitly.
-var compactVerboseObjectFields = map[string]bool{
-	"description": true,
-	"comments":    true,
-	"attachments": true,
-}
-
-// compactFields keeps only the most important fields for agent consumption.
-// For arrays: allowlist of high-gravity fields (no descriptions).
-// For single objects: blocklist that strips known-verbose fields (descriptions, comments, etc.).
-func compactFields(data json.RawMessage) json.RawMessage {
-	// Try array first
-	var items []map[string]any
-	if err := json.Unmarshal(data, &items); err == nil {
-		return compactListFields(items)
-	}
-
-	// Single object — use blocklist
-	var obj map[string]any
-	if err := json.Unmarshal(data, &obj); err == nil {
-		return compactObjectFields(obj)
-	}
-
-	return data
-}
-
-// compactListFields keeps only high-gravity fields for array responses.
-//
-// Two-layer keep rule:
-//
-//  1. A static allow-list covers canonical scalars (id/name/price/status/...).
-//  2. A data-driven extension also keeps any scalar key present in at least
-//     80% of input rows. This catches hand-written novel commands whose
-//     output keys (object_name, match_key, snippet) aren't on the canonical
-//     allow-list, without forcing every printed CLI to expand the list.
-//
-// Verbose fields (description, body, content, etc.) are excluded from the
-// data-driven extension regardless of frequency, so the compact intent
-// (short identifying values for agent consumption, not full prose) is
-// preserved.
-//
-// When an item still carries none of the keep keys, the original is
-// preserved so `--agent` does not silently emit {} for shapes whose key
-// names are entirely off-canonical.
-func compactListFields(items []map[string]any) json.RawMessage {
-	keepFields := map[string]bool{
-		// Identity
-		"id": true, "name": true, "title": true, "identifier": true,
-		"code": true, "slug": true, "key": true,
-		// Categorization
-		"status": true, "state": true, "type": true, "kind": true, "priority": true,
-		// Communication
-		"url": true, "email": true,
-		// Monetary
-		"price": true, "amount": true, "cost": true, "fare": true,
-		"rate": true, "currency": true,
-		// Metrics
-		"rating": true, "score": true, "count": true,
-		// Locale / geo
-		"language": true, "locale": true, "country": true, "region": true,
-		"city": true, "domain": true,
-		// Temporal
-		"created_at": true, "updated_at": true, "createdAt": true, "updatedAt": true,
-		"date": true,
-		// Versioning
-		"version": true,
-	}
-	if len(items) > 0 {
-		keyCounts := map[string]int{}
-		for _, item := range items {
-			for k, v := range item {
-				if compactVerboseListFields[k] || !isCompactScalar(v) {
-					continue
-				}
-				keyCounts[k]++
-			}
-		}
-		// ceil(len(items) * 0.8) without importing math. Capped at len-1 for
-		// len >= 2 so a single missing row cannot veto a key on small lists
-		// (without the cap, ceil(0.8*n) == n for n in {2,3,4}, which silently
-		// reintroduces the partial-strip bug whenever a heterogeneous 2-4 row
-		// response mixes one allow-list key with novel keys).
-		threshold := (len(items)*4 + 4) / 5
-		if len(items) >= 2 && threshold > len(items)-1 {
-			threshold = len(items) - 1
-		}
-		for k, count := range keyCounts {
-			if count >= threshold {
-				keepFields[k] = true
-			}
-		}
-	}
-
-	filtered := make([]map[string]any, 0, len(items))
-	for _, item := range items {
-		compact := map[string]any{}
-		for k, v := range item {
-			if keepFields[k] {
-				compact[k] = v
-			}
-		}
-		if len(compact) == 0 {
-			compact = item
-		}
-		filtered = append(filtered, compact)
-	}
-	result, _ := json.Marshal(filtered)
-	return result
-}
-
-// isCompactScalar reports whether v is a small primitive (string, number,
-// bool, null) suitable for --compact projection. Objects and arrays are
-// rejected: they almost always represent nested structure that defeats the
-// "short identifying value" intent of --compact, and including them in the
-// data-driven keep set would bloat agent output.
-func isCompactScalar(v any) bool {
-	switch v.(type) {
-	case nil, bool, float64, string:
-		return true
-	default:
-		return false
-	}
-}
-
-// compactObjectFields strips known-verbose metadata fields from single-object
-// responses. The blocklist deliberately excludes "body"/"content"/"html"/
-// "markdown" — those fields are payload on `get` commands and stripping them
-// under `--agent`/`--compact` is a silent loss; agents who want to omit them
-// can pass `--select` to specify only the fields they need.
-func compactObjectFields(obj map[string]any) json.RawMessage {
-	compact := map[string]any{}
-	for k, v := range obj {
-		if !compactVerboseObjectFields[k] {
-			compact[k] = v
-		}
-	}
-	result, _ := json.Marshal(compact)
-	return result
-}
-
-// printCSV renders JSON arrays as CSV with header row.
-func printCSV(w io.Writer, data json.RawMessage) error {
-	var items []map[string]any
-	if err := json.Unmarshal(data, &items); err != nil || len(items) == 0 {
-		// Single object or empty - just print as JSON
-		fmt.Fprintln(w, string(data))
-		return nil
-	}
-	// Collect all keys for header
-	keySet := map[string]bool{}
-	for _, item := range items {
-		for k := range item {
-			keySet[k] = true
-		}
-	}
-	var keys []string
-	for k := range keySet {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-	// Header
-	fmt.Fprintln(w, strings.Join(keys, ","))
-	// Rows
-	for _, item := range items {
-		var vals []string
-		for _, k := range keys {
-			v := item[k]
-			if v == nil {
-				vals = append(vals, "")
-			} else {
-				var s string
-				if f, ok := v.(float64); ok {
-					s = strconv.FormatFloat(f, 'f', -1, 64)
-				} else {
-					s = fmt.Sprintf("%v", v)
-				}
-				if strings.ContainsAny(s, ",\"\n") {
-					s = `"` + strings.ReplaceAll(s, `"`, `""`) + `"`
-				}
-				vals = append(vals, s)
-			}
-		}
-		fmt.Fprintln(w, strings.Join(vals, ","))
-	}
-	return nil
-}
-
-// printOutput auto-detects arrays and renders as tables, or prints raw JSON for objects.
-func printOutput(w io.Writer, data json.RawMessage, asJSON bool) error {
-	if !asJSON && !isTerminal(w) {
-		asJSON = true
-	}
-
-	if asJSON {
-		enc := json.NewEncoder(w)
-		enc.SetIndent("", "  ")
-		return enc.Encode(data)
-	}
-
-	// Try to detect if response is an array
-	var items []map[string]any
-	if err := json.Unmarshal(data, &items); err == nil && len(items) > 0 {
-		if err := printAutoTable(w, items); err != nil {
-			return err
-		}
-		// Agent-friendly: show count and suggest narrowing when results are large
-		if len(items) >= 25 {
-			fmt.Fprintf(os.Stderr, "\nShowing %d results. To narrow: add --limit, --json --select, or filter flags.\n", len(items))
-		}
-		return nil
-	}
-
-	// Single object - pretty print
-	var obj map[string]any
-	if err := json.Unmarshal(data, &obj); err == nil {
-		enc := json.NewEncoder(w)
-		enc.SetIndent("", "  ")
-		return enc.Encode(obj)
-	}
-
-	// Fallback: print raw
-	fmt.Fprintln(w, string(data))
-	return nil
-}
-
-// levenshteinDistance computes the edit distance between two strings using a two-row DP approach.
-func levenshteinDistance(a, b string) int {
-	if len(a) == 0 {
-		return len(b)
-	}
-	if len(b) == 0 {
-		return len(a)
-	}
-	if len(a) < len(b) {
-		a, b = b, a
-	}
-	prev := make([]int, len(b)+1)
-	curr := make([]int, len(b)+1)
-	for j := range prev {
-		prev[j] = j
-	}
-	for i := 1; i <= len(a); i++ {
-		curr[0] = i
-		for j := 1; j <= len(b); j++ {
-			cost := 1
-			if a[i-1] == b[j-1] {
-				cost = 0
-			}
-			ins := curr[j-1] + 1
-			del := prev[j] + 1
-			sub := prev[j-1] + cost
-			min := ins
-			if del < min {
-				min = del
-			}
-			if sub < min {
-				min = sub
-			}
-			curr[j] = min
-		}
-		prev, curr = curr, prev
-	}
-	return prev[len(b)]
-}
-
-// suggestFlag returns the closest known flag name to the unknown string, or "" if none is close enough.
-func suggestFlag(unknown string, cmd *cobra.Command) string {
-	unknown = strings.TrimLeft(unknown, "-")
-	best := ""
-	bestDist := 4 // only consider distance <= 3
-	check := func(name string) {
-		d := levenshteinDistance(unknown, name)
-		if d < bestDist && d*5 <= len(unknown)*2 {
-			bestDist = d
-			best = name
-		}
-	}
-	cmd.Flags().VisitAll(func(f *pflag.Flag) {
-		check(f.Name)
-	})
-	cmd.InheritedFlags().VisitAll(func(f *pflag.Flag) {
-		check(f.Name)
-	})
-	return best
-}
-
-// wantsHumanTable returns true when output should be a human-friendly table.
-// Smart default: terminal=table, pipe=JSON.
-// - Human in terminal: isTerminal()=true → table
-// - Claude Code/Codex bash tool: stdout piped → JSON
-// - --json/--csv/--compact/--agent: machine format → JSON
-func wantsHumanTable(w io.Writer, flags *rootFlags) bool {
-	if flags.asJSON || flags.csv || flags.compact || flags.quiet || flags.plain {
-		return false
-	}
-	if flags.selectFields != "" {
-		return false
-	}
-	return isTerminal(w)
-}
-
-func printAutoTable(w io.Writer, items []map[string]any) error {
-	if len(items) == 0 {
-		return nil
-	}
-
-	// Count scalar vs complex fields to decide format
-	scalarCount := 0
-	for _, v := range items[0] {
-		if isCompactScalar(v) {
-			scalarCount++
-		}
-	}
-
-	// Use sectional/card layout for complex items (many fields or nested data)
-	if len(items[0]) > 8 || scalarCount < len(items[0])-2 {
-		return printAutoCards(w, items)
-	}
-
-	headers := prioritizeHeaders(items[0])
-
-	// Limit to 6 columns max for readability
-	if len(headers) > 6 {
-		headers = headers[:6]
-	}
-
-	// Build rows
-	rows := make([][]string, 0, len(items))
-	for _, item := range items {
-		row := make([]string, len(headers))
-		for i, h := range headers {
-			row[i] = formatCellValue(item[h])
-		}
-		rows = append(rows, row)
-	}
-
-	// Print with tab alignment using tabwriter
-	tw := newTabWriter(w)
-	upperHeaders := make([]string, len(headers))
-	for i, h := range headers {
-		upperHeaders[i] = bold(strings.ToUpper(h))
-	}
-
-	fmt.Fprintln(tw, strings.Join(upperHeaders, "\t"))
-	for _, row := range rows {
-		fmt.Fprintln(tw, strings.Join(row, "\t"))
-	}
-	return tw.Flush()
-}
-
-// prioritizeHeaders orders scalar fields by importance for table display.
-func prioritizeHeaders(item map[string]any) []string {
-	return prioritizeFields(item, false)
-}
-
-// prioritizeAllHeaders orders all fields (including arrays) by importance for card display.
-func prioritizeAllHeaders(item map[string]any) []string {
-	return prioritizeFields(item, true)
-}
-
-// prioritizeFields orders fields by importance: identity → temporal → status → other.
-// When includeComplex is true, arrays and objects are included (for card layout).
-//
-// Uses exact-or-suffix matching to avoid false positives: "name" matches "Name" and
-// "UserName" but not "BuildingName" (because "Building" is not a known prefix that
-// indicates identity). The field is split on camelCase/snake_case boundaries and the
-// LAST segment is matched against patterns.
-func prioritizeFields(item map[string]any, includeComplex bool) []string {
-	// Priority tiers — matched against the last segment of the field name.
-	// "OrderDate" → last segment "date" → tier 1 (temporal).
-	// "BuildingName" → last segment "name" → tier 0... but we want to avoid this.
-	// Solution: exact match on the full lowered name OR suffix segment match,
-	// with a penalty for compound names that have a non-identity prefix.
-	type pattern struct {
-		word string
-		tier int
-	}
-	// Exact matches (full field name, case-insensitive) — highest confidence
-	exactMatches := map[string]int{
-		"id": 0, "name": 0, "title": 0, "slug": 0, "key": 0,
-		"date": 1, "created": 1, "updated": 1, "createdat": 1, "updatedat": 1,
-		"status": 2, "state": 2, "statuscode": 2,
-		"summary": 3, "description": 3, "price": 3, "amount": 3, "total": 3,
-		"cost": 3, "points": 3, "score": 3,
-		"type": 4, "kind": 4, "category": 4, "email": 4, "phone": 4, "url": 4,
-	}
-	// Suffix patterns — match when the field ends with this word (after splitting)
-	suffixMatches := map[string]int{
-		"id": 0, "name": 0, "title": 0,
-		"date": 1, "time": 1,
-		"status": 2, "state": 2, "code": 2,
-		"price": 3, "amount": 3, "total": 3, "cost": 3,
-		"summary": 3, "description": 3, "points": 3, "score": 3,
-		"type": 4, "kind": 4, "category": 4, "method": 4,
-	}
-
-	numTiers := 5
-
-	type scored struct {
-		name  string
-		tier  int
-		index int
-	}
-
-	var all []scored
-	idx := 0
-	for k, v := range item {
-		if !includeComplex {
-			switch v.(type) {
-			case []any, map[string]any:
-				continue
-			}
-		}
-		// Skip values that won't render usefully in cards
-		if includeComplex {
-			formatted := formatCellValue(v)
-			if formatted == "" {
-				continue
-			}
-		}
-
-		tier := numTiers // default: unclassified
-		lower := strings.ToLower(k)
-
-		// 1. Exact match on full field name
-		if t, ok := exactMatches[lower]; ok {
-			tier = t
-		} else {
-			// 2. Split camelCase into segments and match the last one
-			segments := splitCamelCase(lower)
-			if len(segments) > 0 {
-				lastSeg := segments[len(segments)-1]
-				if t, ok := suffixMatches[lastSeg]; ok {
-					// Compound names with identity suffixes (BuildingName, TipTime)
-					// get demoted one tier because the prefix dilutes the signal
-					if len(segments) > 1 {
-						tier = t + 1
-					} else {
-						tier = t
-					}
-				}
-			}
-		}
-
-		// Demote booleans to last
-		if _, ok := v.(bool); ok && tier >= numTiers {
-			tier = numTiers + 1
-		}
-		all = append(all, scored{name: k, tier: tier, index: idx})
-		idx++
-	}
-
-	sort.Slice(all, func(i, j int) bool {
-		if all[i].tier != all[j].tier {
-			return all[i].tier < all[j].tier
-		}
-		return all[i].index < all[j].index
-	})
-
-	headers := make([]string, len(all))
-	for i, s := range all {
-		headers[i] = s.name
-	}
-	return headers
-}
-
-// splitCamelCase splits "OrderDate" → ["order", "date"], "statusCode" → ["status", "code"],
-// "page_size" → ["page", "size"].
-func splitCamelCase(s string) []string {
-	var segments []string
-	var current strings.Builder
-	runes := []rune(s)
-	for i, r := range runes {
-		if r == '_' || r == '-' {
-			if current.Len() > 0 {
-				segments = append(segments, current.String())
-				current.Reset()
-			}
-			continue
-		}
-		if i > 0 && unicode.IsUpper(r) && unicode.IsLower(runes[i-1]) {
-			if current.Len() > 0 {
-				segments = append(segments, current.String())
-				current.Reset()
-			}
-		}
-		current.WriteRune(unicode.ToLower(r))
-	}
-	if current.Len() > 0 {
-		segments = append(segments, current.String())
-	}
-	return segments
-}
-
-// printAutoCards renders items as labeled cards — one block per item.
-// Used for complex responses with many fields or nested data.
-func printAutoCards(w io.Writer, items []map[string]any) error {
-	headers := prioritizeAllHeaders(items[0])
-
-	// Find the longest header for alignment (from fields we'll actually show)
-	maxLen := 0
-	for _, h := range headers {
-		if len(h) > maxLen {
-			maxLen = len(h)
-		}
-	}
-
-	for i, item := range items {
-		if i > 0 {
-			fmt.Fprintln(w)
-		}
-
-		// Card header: use first priority field as the card title
-		titleVal := formatCellValue(item[headers[0]])
-		if len(headers) > 1 {
-			secondVal := formatCellValue(item[headers[1]])
-			if secondVal != "" {
-				fmt.Fprintf(w, "%s %s — %s\n", bold(strings.ToUpper(headers[0])), titleVal, secondVal)
-			} else {
-				fmt.Fprintf(w, "%s %s\n", bold(strings.ToUpper(headers[0])), titleVal)
-			}
-		} else {
-			fmt.Fprintf(w, "%s %s\n", bold(strings.ToUpper(headers[0])), titleVal)
-		}
-
-		// Remaining fields indented — skip empty, zero, and false values
-		for _, h := range headers[2:] {
-			v := formatCellValue(item[h])
-			if v == "" || v == "false" || v == "0" || v == "[]" || v == "null" {
-				continue
-			}
-			// Multi-line values (nested arrays) start with \n
-			if strings.HasPrefix(v, "\n") {
-				fmt.Fprintf(w, "  %s:%s\n", h, v)
-			} else {
-				fmt.Fprintf(w, "  %-*s  %s\n", maxLen, h+":", v)
-			}
-		}
-	}
-	return nil
-}
-
-func formatCellValue(v any) string {
-	switch val := v.(type) {
-	case string:
-		// Format ISO dates as just the date portion
-		if len(val) >= 19 && val[4] == '-' && val[7] == '-' && val[10] == 'T' {
-			return val[:10]
-		}
-		return truncate(val, 60)
-	case float64:
-		if val == float64(int64(val)) {
-			return fmt.Sprintf("%d", int64(val))
-		}
-		return fmt.Sprintf("%.2f", val)
-	case bool:
-		return fmt.Sprintf("%t", val)
-	case nil:
-		return ""
-	case []any:
-		if len(val) == 0 {
-			return ""
-		}
-		// If array contains objects, format each as a summary line
-		if obj, isObj := val[0].(map[string]any); isObj {
-			_ = obj
-			return formatObjectArray(val)
-		}
-		// Flatten simple arrays into comma-separated string
-		parts := make([]string, 0, len(val))
-		for _, item := range val {
-			if s, ok := item.(string); ok {
-				parts = append(parts, s)
-			} else {
-				b, _ := json.Marshal(item)
-				parts = append(parts, string(b))
-			}
-		}
-		return truncate(strings.Join(parts, ", "), 60)
-	case map[string]any:
-		return formatSingleObject(val)
-	default:
-		b, _ := json.Marshal(val)
-		return truncate(string(b), 60)
-	}
-}
-
-// formatObjectArray renders an array of objects as multi-line summary.
-// Each object is summarized by its most descriptive fields: name/title, qty, size, price.
-func formatObjectArray(items []any) string {
-	var lines []string
-	for _, raw := range items {
-		obj, ok := raw.(map[string]any)
-		if !ok {
-			continue
-		}
-		lines = append(lines, formatObjectSummary(obj))
-	}
-	if len(lines) == 0 {
-		return ""
-	}
-	// Multi-line: newline-prefixed so the card renderer can indent
-	return "\n" + strings.Join(lines, "\n")
-}
-
-// formatObjectSummary extracts the most useful fields from an object into a one-line summary.
-// Looks for: qty/count → name/title → size → price, in that order.
-func formatObjectSummary(obj map[string]any) string {
-	var parts []string
-
-	// Quantity
-	qty := findField(obj, "qty", "count", "quantity")
-	if qty != "" && qty != "1" && qty != "0" {
-		parts = append(parts, qty+"x")
-	} else if qty == "1" {
-		parts = append(parts, "1x")
-	}
-
-	// Name — check nested objects too (e.g., Side1.Name)
-	name := findField(obj, "name", "title", "label", "description")
-	if name == "" {
-		// Check nested objects for name
-		for _, key := range []string{"Side1", "side1", "Item", "item", "Product", "product"} {
-			if nested, ok := obj[key].(map[string]any); ok {
-				name = findField(nested, "name", "title", "label")
-				if name != "" {
-					break
-				}
-			}
-		}
-	}
-	if name != "" {
-		parts = append(parts, name)
-	}
-
-	// Size
-	size := findField(obj, "sizename", "size_name")
-	if size == "" {
-		size = findField(obj, "catname", "cat_name", "category")
-	}
-	if size != "" {
-		parts = append(parts, "—")
-		parts = append(parts, size)
-	}
-
-	// Price
-	price := findField(obj, "extprice", "price", "amount", "total")
-	if price != "" && price != "0" {
-		parts = append(parts, fmt.Sprintf("($%s)", price))
-	}
-
-	if len(parts) == 0 {
-		// Fallback: JSON summary
-		b, _ := json.Marshal(obj)
-		return truncate(string(b), 80)
-	}
-	return "    " + strings.Join(parts, " ")
-}
-
-// formatSingleObject renders a single object by its most descriptive fields.
-func formatSingleObject(obj map[string]any) string {
-	name := findField(obj, "name", "title", "label", "description")
-	if name != "" {
-		return name
-	}
-	id := findField(obj, "id", "key", "code")
-	if id != "" {
-		return id
-	}
-	return ""
-}
-
-// findField searches an object for a field name (case-insensitive) and returns its formatted value.
-func findField(obj map[string]any, names ...string) string {
-	for _, name := range names {
-		for k, v := range obj {
-			if strings.EqualFold(k, name) {
-				return formatCellValue(v)
-			}
-		}
-	}
-	return ""
 }
